move uri-util to core
