+++ core/package.json
@@ -26,7 +26,9 @@
     "url": "https://github.com/neo4j/neo4j-javascript-driver/issues"
   },
   "homepage": "https://github.com/neo4j/neo4j-javascript-driver#readme",
-  "dependencies": {},
+  "dependencies": {
+    "uri-js": "^4.4.1"
+  },
   "devDependencies": {
     "@types/jest": "^26.0.20",
     "esdoc": "^1.1.0",
